pyproject.toml
[project]
name = "project-01-rakuten-multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "kaggle>=2.0.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.16.0",
    "matplotlib>=3.8.0",
    "nltk>=3.9.4",
    "torch",
    "torchvision",
    "torchaudio",
    "seaborn>=0.13.2",
    "transformers>=5.5.4",
    "sentencepiece>=0.2.1",
    "protobuf>=7.34.0",
    "huggingface-hub>=1.11.0",
    "opencv-python>=4.13.0.92",
    "timm>=1.0.26",
    "grad-cam>=1.5.5",
    "streamlit>=1.40.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" }
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" }
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" }
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
